+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Features
 
+- Add `isiOSAppOnMac` and `isMacCatalystApp` from ProcessInfo to the runtime context (#5570)
 - Add experimental support for capturing structured logs via `SentrySDK.logger` (#5532, #5593, #5639, #5628, #5637)
 - The SDK will show a warning in the console if it detects it was loaded twice (#5298)